add responseEntity for all controller and therefore adapt some test

+++ backend-cible/src/main/java/com/cible/backend_cible/service/task/PrioritySERVICE.java
@@ -30,7 +30,12 @@ public class PrioritySERVICE {
         return priorityDB.save(priority);
     }
 
-    public void deletePriority(Integer id) {
+    public boolean deletePriority(Integer id) {
+        if (!priorityDB.existsById(id)) {
+            return false;
+        }
         priorityDB.deleteById(id);
+        return true;
     }
+    
 }
